requirements.txt
fastapi
uvicorn
line-bot-sdk>=3.0.0
openai>=1.10.0
supabase-py-async
python-dotenv